chore: add inspect.ts and rename paginated message
