+++ src/lib/setup/index.ts
@@ -3,7 +3,8 @@ import '#root/config';
 import 'reflect-metadata';
 
 // Import everything else:
-import '#lib/setup/PaginatedMessage';
+import '#lib/setup/paginated-message';
+import '#lib/setup/inspect';
 import '#utils/Sanitizer/initClean';
 import '@sapphire/plugin-api/register';
 import '@sapphire/plugin-editable-commands/register';
@@ -12,8 +13,6 @@ import '@sapphire/plugin-logger/register';
 
 import { ApplicationCommandRegistries, RegisterBehavior } from '@sapphire/framework';
 import * as colorette from 'colorette';
-import { inspect } from 'node:util';
 
-inspect.defaultOptions.depth = 1;
 colorette.createColors({ useColor: true });
 ApplicationCommandRegistries.setDefaultBehaviorWhenNotIdentical(RegisterBehavior.Overwrite);
