+++ src/lib/setup/inspect.ts
@@ -1,0 +1,4 @@
+import { inspect } from 'node:util';
+
+inspect.defaultOptions.depth = 1;
+inspect.defaultOptions.breakLength = 120;
